Make setting up env optional and improve docs

+++ tests/benchmarks/archer2/run_regression_test.sh
@@ -8,6 +8,7 @@ END_REF=""
 START_VENV=".venv-start"
 END_VENV=".venv-end"
 ACCOUNT=""
+SETUP_ENVS="true"
 
 help() {
   echo "Usage:"
@@ -20,6 +21,9 @@ help() {
   echo "    -e | --end-ref <end_ref>         The git ref to be used as the final state."
   echo "    -a | --account <archer2_account> The archer2 account code to run jobs against."
   echo "    -g | --gpu                       Flag to state the gpu benchmark should be ran."
+  echo "    --skip-setup                     Flag to state if the setup (installation of "
+  echo "                                     dependencies) can be skipped. Good for simply"
+  echo "                                     re-submitting"
 }
 
 # check for no input arguments and show help
@@ -61,6 +65,11 @@ while [ $# -gt 0 ] ; do
             shift 1
             continue
             ;;
+        --skip-setup)
+            SETUP_ENVS="false"
+            shift 1
+            continue
+            ;;
         *)
             echo "Invalid option: $1" >&2;
             help
@@ -93,32 +102,37 @@ fi
 
 BENCHMARKS_DIR="$GLASS_DIR/tests/benchmarks"
 
-# Setup base environment
-rm -rf "${GLASS_DIR:?}/$START_VENV" # Cleanup old venv
-uv venv  "$GLASS_DIR/$START_VENV"
-source "$GLASS_DIR/$START_VENV/bin/activate"
-if [ $CPU_OR_GPU = "gpu" ]; then
-  uv sync --active --group test --group archer2-gpu
-else
-  uv sync --active --group test
-fi
-uv pip uninstall glass -y # Make sure no installation of glass already exists
-uv pip install "git+$GLASS_REPO_URL@$START_REF"
 
-# Setup head environment
-rm -rf "${GLASS_DIR:?}/$END_VENV" # Cleanup old venv
-uv venv "$GLASS_DIR/$END_VENV"
-source "$GLASS_DIR/$END_VENV/bin/activate"
-if [ $CPU_OR_GPU = "gpu" ]; then
-  uv sync --active --group test --group archer2-gpu
-else
-  uv sync --active --group test
-fi
-uv pip uninstall glass -y # Make sure no installation of glass already exists
-uv pip install "git+$GLASS_REPO_URL@$END_REF"
+if [[ "$SETUP_ENVS" == "true" ]]
+then
+  # Setup base environment
+  rm -rf "${GLASS_DIR:?}/$START_VENV" # Cleanup old venv
+  uv venv  "$GLASS_DIR/$START_VENV"
+  source "$GLASS_DIR/$START_VENV/bin/activate"
+  if [ $CPU_OR_GPU = "gpu" ]; then
+    uv sync --active --group test --group archer2-gpu
+  else
+    uv sync --active --group test
+  fi
+  uv pip uninstall glass -y # Make sure no installation of glass already exists
+  uv pip install "git+$GLASS_REPO_URL@$START_REF"
 
-# Remove old benchmark results
-rm -rf "$BENCHMARKS_DIR/outputs"
+  # Setup head environment
+  rm -rf "${GLASS_DIR:?}/$END_VENV" # Cleanup old venv
+  uv venv "$GLASS_DIR/$END_VENV"
+  source "$GLASS_DIR/$END_VENV/bin/activate"
+  if [ $CPU_OR_GPU = "gpu" ]; then
+    uv sync --active --group test --group archer2-gpu
+  else
+    uv sync --active --group test
+  fi
+  uv pip uninstall glass -y # Make sure no installation of glass already exists
+  uv pip install "git+$GLASS_REPO_URL@$END_REF"
+
+  # Remove old benchmark results
+  rm -rf "$BENCHMARKS_DIR/outputs"
+fi
 
 # Submit job
-sbatch --account="$ACCOUNT" "$BENCHMARKS_DIR/archer2/submission_script_$CPU_OR_GPU.sh" "$GLASS_DIR"
+sbatch --account="$ACCOUNT" --output="%x-$START_REF-$END_REF-%j.out" \
+  "$BENCHMARKS_DIR/archer2/submission_script_$CPU_OR_GPU.sh" "$GLASS_DIR"
